requirements.txt
certifi==2025.1.31
funasr==1.2.6
numpy==2.2.4
openai==1.68.2
PyAudio==0.2.14
PyYAML==6.0.2
Requests==2.32.3
tencentcloud_sdk_python==3.0.1345
websockets==15.0.1
